Use explicit casts for size_t to unsigned conversion https://bugs.webkit.org/show_bug.cgi?id=83602
Reviewed by Eric Seidel.

No new tests, no change in functinality.

Not all platforms have implicit casts from size_t to unsigned and we
can't add size_t versions of the operators to FractionalLayoutUnit for
all platforms as it would conflict with the unsigned versions of same.

Change support methods to take unsigned instead of size_t and use
explicit casts when multiplying or dividing a FractionalLayoutUnit with
a size_t and the other way around.

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::adjustForColumns):
* rendering/RenderFlexibleBox.cpp:
(WebCore::initialPackingOffset):
(WebCore::packingSpaceBetweenChildren):
(WebCore::initialLinePackingOffset):
(WebCore::linePackingSpaceBetweenChildren):
(WebCore::RenderFlexibleBox::packFlexLines):
* rendering/RenderFrameSet.cpp:
(WebCore::RenderFrameSet::layout):


Canonical link: https://commits.webkit.org/101772@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114537 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/rendering/RenderBlock.cpp
@@ -5176,11 +5176,11 @@ void RenderBlock::adjustForColumns(LayoutSize& offset, const LayoutPoint& point)
     ColumnInfo* colInfo = columnInfo();
 
     LayoutUnit logicalLeft = logicalLeftOffsetForContent();
-    size_t colCount = columnCount(colInfo);
+    unsigned colCount = columnCount(colInfo);
     LayoutUnit colLogicalWidth = colInfo->desiredColumnWidth();
     LayoutUnit colLogicalHeight = colInfo->columnHeight();
 
-    for (size_t i = 0; i < colCount; ++i) {
+    for (unsigned i = 0; i < colCount; ++i) {
         // Compute the edges for a given column in the block progression direction.
         LayoutRect sliceRect = LayoutRect(logicalLeft, borderBefore() + paddingBefore() + i * colLogicalHeight, colLogicalWidth, colLogicalHeight);
         if (!isHorizontalWritingMode())
